requirements_dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements_dev.txt requirements_dev.in
#
asgiref==3.5.2
    # via
    #   -r ./requirements.txt
    #   django
attrs==21.4.0
    # via
    #   -r ./requirements.txt
    #   jsonschema
    #   pytest
black==22.6.0
    # via -r requirements_dev.in
build==0.8.0
    # via pip-tools
certifi==2022.6.15
    # via
    #   -r ./requirements.txt
    #   requests
cffi==1.15.1
    # via
    #   -r ./requirements.txt
    #   cryptography
charset-normalizer==2.1.0
    # via
    #   -r ./requirements.txt
    #   requests
click==8.1.3
    # via
    #   -r ./requirements.txt
    #   black
    #   flex
    #   pip-tools
coreapi==2.3.3
    # via
    #   -r ./requirements.txt
    #   drf-yasg
coreschema==0.0.4
    # via
    #   -r ./requirements.txt
    #   coreapi
    #   drf-yasg
coverage[toml]==6.4.2
    # via pytest-cov
cryptography==37.0.4
    # via
    #   -r ./requirements.txt
    #   josepy
    #   mozilla-django-oidc
    #   pyopenssl
datapunt-keycloak-oidc==0.5.1
    # via -r ./requirements.txt
defusedxml==0.7.1
    # via
    #   -r ./requirements.txt
    #   djangorestframework-xml
django==3.2.14
    # via
    #   -r ./requirements.txt
    #   django-braces
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   django-extra-fields
    #   django-filter
    #   django-leaflet
    #   django-multiselectfield
    #   djangorestframework
    #   drf-yasg
    #   mozilla-django-oidc
django-braces==1.15.0
    # via -r ./requirements.txt
django-cors-headers==3.13.0
    # via -r ./requirements.txt
django-countries==7.3.2
    # via -r ./requirements.txt
django-debug-toolbar==3.5.0
    # via -r requirements_dev.in
django-extensions==3.2.0
    # via -r ./requirements.txt
django-extra-fields==3.0.2
    # via -r ./requirements.txt
django-filter==22.1
    # via -r ./requirements.txt
django-leaflet==0.28.3
    # via -r ./requirements.txt
django-multiselectfield==0.1.12
    # via -r ./requirements.txt
djangorestframework==3.13.1
    # via
    #   -r ./requirements.txt
    #   django-extra-fields
    #   djangorestframework-csv
    #   djangorestframework-gis
    #   drf-amsterdam
    #   drf-extensions
    #   drf-yasg
djangorestframework-csv==2.1.1
    # via
    #   -r ./requirements.txt
    #   drf-amsterdam
djangorestframework-gis==1.0
    # via -r ./requirements.txt
djangorestframework-xml==2.0.0
    # via
    #   -r ./requirements.txt
    #   drf-amsterdam
drf-amsterdam==0.4.1
    # via -r ./requirements.txt
drf-extensions==0.7.1
    # via
    #   -r ./requirements.txt
    #   drf-amsterdam
drf-yasg==1.21.0
    # via -r ./requirements.txt
et-xmlfile==1.1.0
    # via
    #   -r ./requirements.txt
    #   openpyxl
factory-boy==3.2.1
    # via
    #   -r requirements_dev.in
    #   pytest-factoryboy
faker==13.15.0
    # via
    #   -r requirements_dev.in
    #   factory-boy
    #   pytest-faker
flex==6.14.1
    # via -r ./requirements.txt
graypy==2.1.0
    # via -r ./requirements.txt
idna==3.3
    # via
    #   -r ./requirements.txt
    #   requests
importlib-metadata==4.12.0
    # via
    #   -r ./requirements.txt
    #   markdown
importlib-resources==5.8.0
    # via
    #   -r ./requirements.txt
    #   jsonschema
inflection==0.5.1
    # via
    #   -r ./requirements.txt
    #   drf-yasg
    #   pytest-factoryboy
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via -r requirements_dev.in
itypes==1.2.0
    # via
    #   -r ./requirements.txt
    #   coreapi
jinja2==3.1.2
    # via
    #   -r ./requirements.txt
    #   coreschema
josepy==1.13.0
    # via
    #   -r ./requirements.txt
    #   mozilla-django-oidc
jsonpointer==2.3
    # via
    #   -r ./requirements.txt
    #   flex
jsonschema==4.7.2
    # via
    #   -r ./requirements.txt
    #   swagger-spec-validator
markdown==3.3.7
    # via -r ./requirements.txt
markupsafe==2.1.1
    # via
    #   -r ./requirements.txt
    #   jinja2
mozilla-django-oidc==1.2.4
    # via
    #   -r ./requirements.txt
    #   datapunt-keycloak-oidc
mypy-extensions==0.4.3
    # via black
openpyxl==3.0.10
    # via -r ./requirements.txt
packaging==21.3
    # via
    #   -r ./requirements.txt
    #   build
    #   drf-yasg
    #   pytest
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via build
pip-tools==6.8.0
    # via -r requirements_dev.in
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via pytest
psycopg2-binary==2.9.3
    # via -r ./requirements.txt
py==1.11.0
    # via pytest
pycparser==2.21
    # via
    #   -r ./requirements.txt
    #   cffi
pygelf==0.4.2
    # via -r ./requirements.txt
pygments==2.12.0
    # via -r ./requirements.txt
pyinstrument==4.2.0
    # via -r requirements_dev.in
pyopenssl==22.0.0
    # via
    #   -r ./requirements.txt
    #   josepy
pyparsing==3.0.9
    # via
    #   -r ./requirements.txt
    #   packaging
pyrsistent==0.18.1
    # via
    #   -r ./requirements.txt
    #   jsonschema
pytest==7.1.2
    # via
    #   -r requirements_dev.in
    #   pytest-cov
    #   pytest-django
    #   pytest-env
    #   pytest-factoryboy
pytest-cov==3.0.0
    # via -r requirements_dev.in
pytest-django==4.5.2
    # via -r requirements_dev.in
pytest-env==0.6.2
    # via -r requirements_dev.in
pytest-factoryboy==2.5.0
    # via -r requirements_dev.in
pytest-faker==2.0.0
    # via -r requirements_dev.in
python-dateutil==2.8.2
    # via faker
pytz==2022.1
    # via
    #   -r ./requirements.txt
    #   django
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0
    # via
    #   -r ./requirements.txt
    #   flex
    #   swagger-spec-validator
raven==6.10.0
    # via -r ./requirements.txt
requests==2.28.1
    # via
    #   -r ./requirements.txt
    #   coreapi
    #   flex
    #   mozilla-django-oidc
    #   responses
responses==0.21.0
    # via -r requirements_dev.in
rfc3987==1.3.8
    # via
    #   -r ./requirements.txt
    #   flex
ruamel-yaml==0.17.21
    # via
    #   -r ./requirements.txt
    #   drf-yasg
ruamel-yaml-clib==0.2.6
    # via
    #   -r ./requirements.txt
    #   ruamel-yaml
six==1.16.0
    # via
    #   -r ./requirements.txt
    #   djangorestframework-csv
    #   flex
    #   mozilla-django-oidc
    #   python-dateutil
    #   swagger-spec-validator
sqlparse==0.4.2
    # via
    #   -r ./requirements.txt
    #   django
    #   django-debug-toolbar
strict-rfc3339==0.7
    # via
    #   -r ./requirements.txt
    #   flex
swagger-spec-validator==2.7.4
    # via -r ./requirements.txt
tomli==2.0.1
    # via
    #   black
    #   build
    #   coverage
    #   pytest
typing-extensions==4.3.0
    # via
    #   -r ./requirements.txt
    #   black
    #   django-countries
    #   pytest-factoryboy
unicodecsv==0.14.1
    # via
    #   -r ./requirements.txt
    #   djangorestframework-csv
uritemplate==4.1.1
    # via
    #   -r ./requirements.txt
    #   coreapi
    #   drf-yasg
urllib3==1.26.10
    # via
    #   -r ./requirements.txt
    #   requests
    #   responses
validate-email==1.3
    # via
    #   -r ./requirements.txt
    #   flex
wheel==0.37.1
    # via pip-tools
zipp==3.8.1
    # via
    #   -r ./requirements.txt
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
